Refactor duplicated `validate_gossiped_execution_receipt()`
An easy obvious refactoring could be done, there is a great chance
the `core_gossip_message_validator` and `system_gossip_message_validator`
can be refined further, but I'll skip it for now as it's not the goal of
this PR.
